LayoutTests/sputnik/Conformance/12_Statement/12.6_Iteration_Statements/12.6.3_The_for_Statement/S12.6.3_A4_T1-expected.txt

CONSOLE MESSAGE: line 80: SyntaxError: Expected token ')'
S12.6.3_A4_T1

PASS Expected parsing failure

TEST COMPLETE
